handle auth check & handle post user relations

+++ backend/controllers/feed.js
@@ -36,7 +36,7 @@ exports.getPosts = (req, res, next) => {
 exports.createPost = (req, res, next) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
-    const error = new Error("Validation failed, entered data is incorrect");
+    const error = new Error("Validation failed, entered data is incorrect.");
     error.statusCode = 422;
     throw error;
   }
@@ -45,7 +45,6 @@ exports.createPost = (req, res, next) => {
     error.statusCode = 422;
     throw error;
   }
-
   const imageUrl = req.file.path.replace(/\\/g, "/");
   const title = req.body.title;
   const content = req.body.content;
@@ -68,9 +67,12 @@ exports.createPost = (req, res, next) => {
     })
     .then(() => {
       res.status(201).json({
-        message: "Post created successfully",
+        message: "Post created successfully!",
         post: post,
-        creator: { _id: creator._id, name: creator.name },
+        creator: {
+          _id: creator._id,
+          name: creator.name,
+        },
       });
     })
     .catch((err) => {
@@ -90,7 +92,10 @@ exports.getPost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-      res.status(200).json({ message: "Post fetched", post: post });
+      res.status(200).json({
+        message: "Post fetched.",
+        post: post,
+      });
     })
     .catch((err) => {
       if (!err.statusCode) {
@@ -112,6 +117,7 @@ exports.updatePost = (req, res, next) => {
   const content = req.body.content;
   let imageUrl = req.body.image;
   if (req.file) {
+    /** REPLACE ALL '\' WITH '/' */
     imageUrl = req.file.path.replace(/\\/g, "/");
   }
   if (!imageUrl) {
@@ -124,6 +130,11 @@ exports.updatePost = (req, res, next) => {
       if (!post) {
         const error = new Error("Could not find post.");
         error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized.");
+        error.statusCode = 403;
         throw error;
       }
       if (imageUrl !== post.imageUrl) {
@@ -154,11 +165,24 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized.");
+        error.statusCode = 403;
+        throw error;
+      }
+      // Check logged in user
       clearImage(post.imageUrl);
+      /** .findByIdAndRemove -- REMOVED */
       return Post.findByIdAndDelete(postId);
     })
-    .then((result) => {
-      console.log(result);
+    .then(() => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(() => {
       res.status(200).json({ message: "Deleted post." });
     })
     .catch((err) => {
